Make launcher and block selectable and do not deselect when click on panel
By default clicking on panel deselects previous selection. This is solved by making GameManager storing the Selectable of last selected launcher/block.
Launcher now rotates to face the touch position.
Launching is still problematic. The issue lies in preventing the launcher to launch right after creating it or finish dragging it. This is an issue because launching is handled by SimulationPanel, which calls launcher (stored in GameManager) to launch when it deems appropriate by checking TouchPhase.