Update README.md
At the line 
"For Android users, find the APK and download it here.",
The hyperlink at "here" is linked to the root directory. 
Though the APK is indeed at the root directory, we did not explicitly mention that.
This has confused some users.

Now, the hyperlink targets the Milestone3APK3.apk file itself.

Note that this means we need to update the link for every future releases.